pyproject.toml
[project]
name = "garmin-workout-pipeline"
version = "0.1.0"
description = "YAML to Garmin Connect workout pipeline"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "garminconnect>=0.3",
    "pydantic>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
gwp = "cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["garmin_pipeline"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
